fix build errors

+++ hotspot/src/share/vm/opto/library_call.cpp
@@ -3825,8 +3825,8 @@ Node* LibraryCallKit::generate_array_guard_common(Node* kls, RegionNode* region,
   }
   // Now test the correct condition.
   jint  nval = (obj_array
-                ? ((jint)Klass::_lh_array_tag_type_value
-                   <<    Klass::_lh_array_tag_shift)
+                ? ((jint)((unsigned long)Klass::_lh_array_tag_type_value
+                   <<    Klass::_lh_array_tag_shift))
                 : Klass::_lh_neutral_value);
   Node* cmp = _gvn.transform(new(C) CmpINode(layout_val, intcon(nval)));
   BoolTest::mask btest = BoolTest::lt;  // correct for testing is_[obj]array
